SensorSystemService: Split into Interface and an implementation
